+++ app/Http/Livewire/table/detail/UserDetail.php
@@ -3,18 +3,21 @@
 namespace App\Http\Livewire\table\detail;
 
 use App\Models\User;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Str;
 
 class UserDetail extends Detail
 {
-    public $user;
-    public $changePassword = false;
 
-    protected $rules =[
+    public $user;
+
+    protected $rules = [
         'user.id' => 'sometimes|required|exists:users,id',
         'user.name' => 'required|string',
     ];
 
-    public function mount($object){
+    public function mount($object)
+    {
         $this->user = $object;
     }
 
@@ -25,8 +28,38 @@ class UserDetail extends Detail
 
     public function delete()
     {
-        $user = User::query()->whereKey($this->user['id'])->first();
+        $user = User::query()
+            ->whereKey($this->user['id'])
+            ->first();
+
+        if ($user->parties()->count()) {
+            $this->notification([
+                'title' => 'Can not delete!',
+                'description' => 'User is assigned to at least one party.',
+                'icon' => 'error',
+                'timeout' => 5000,
+            ]);
+            return;
+        }
+
+        if ($user->meals()->count()){
+            $this->notification([
+                'title' => 'Can not delete!',
+                'description' => 'User is assigned to at least one meal.',
+                'icon' => 'error',
+                'timeout' => 5000,
+            ]);
+            return;
+        }
+
         $user->delete();
+
+        $this->notification([
+            'title' => 'User deleted.',
+            'description' => 'User was successful deleted.',
+            'icon' => 'success',
+            'timeout' => 5000,
+        ]);
     }
 
     public function save()
@@ -38,5 +71,14 @@ class UserDetail extends Detail
         $user->save();
         $this->user = $user->toArray();
         $this->inEditState = false;
+    }
+
+    public function resetPassword()
+    {
+        $user = User::query()->whereKey($this->user['id'])->first();
+        $password = Str::random(10);
+        $user->password = Hash::make($password);
+        $user->save();
+        $this->dialog()->info('New password',$password);
     }
 }
